pyproject.toml
[tool.poetry]
name = "pysynccalendar"
version = "1.0.0"
description = ""
authors = ["chanma0806"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.127.0"
google-auth-oauthlib = "^1.2.0"
python-dotenv = "^1.0.1"
pywin32 = "306"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
